Improve mech core handling
Mechs now must always have a core. This is what distinguishes a mech from a cluster. Since they always have a core, Mech.CoreBlockId holds the constant ID of the core block (right now it is 0, which is unlikely to change). You also now cannot add another core block to a mech.